requirements.txt
Django~=3.2.10
python-dotenv~=1.0.0
asgiref==3.4.1
django-filter~=23.4
pip~=23.3.1
wheel~=0.41.3
cryptography~=41.0.7
setuptools~=68.2.2
djoser~=2.2.2
djangorestframework~=3.14.0
Pillow~=10.1.0
django-ckeditor~=6.7.0